Add throws exceptions

+++ src/main/java/by/akimova/CartAPI/service/impl/CartServiceImpl.java
@@ -1,15 +1,17 @@
 package by.akimova.CartAPI.service.impl;
 
 import by.akimova.CartAPI.exception.EntityNotFoundException;
-import by.akimova.CartAPI.exception.ValidationException;
+import by.akimova.CartAPI.exception.NotAccessException;
+import by.akimova.CartAPI.exception.NotValidUsernameException;
 import by.akimova.CartAPI.model.Cart;
 import by.akimova.CartAPI.model.Item;
+import by.akimova.CartAPI.model.User;
 import by.akimova.CartAPI.repository.CartRepository;
 import by.akimova.CartAPI.repository.ItemRepository;
+import by.akimova.CartAPI.repository.UserRepository;
 import by.akimova.CartAPI.service.CartService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
@@ -28,6 +30,7 @@ import java.util.*;
 public class CartServiceImpl implements CartService {
 
     private final CartRepository cartRepository;
+    private final UserRepository userRepository;
     private final ItemRepository itemRepository;
 
     /**
@@ -50,11 +53,11 @@ public class CartServiceImpl implements CartService {
      * @return found cart.
      */
     @Override
-    public Cart getCartById(UUID cartId) throws ValidationException, EntityNotFoundException {
+    public Cart getCartById(UUID cartId) throws NotValidUsernameException, EntityNotFoundException {
 
         if (cartId == null) {
             log.error("IN getCartById - id is null ");
-            throw new ValidationException("cartId is null");
+            throw new NotValidUsernameException("cartId is null");
         }
 
         Cart cart = cartRepository.findCartByCartId(cartId);
@@ -75,11 +78,11 @@ public class CartServiceImpl implements CartService {
      * @return found cart.
      */
     @Override
-    public Cart getCartByUserId(UUID userId) throws EntityNotFoundException, ValidationException {
+    public Cart getCartByUserId(UUID userId) throws EntityNotFoundException, NotValidUsernameException {
 
         if (userId == null) {
             log.error("IN getCartByUserId - id is null ");
-            throw new ValidationException("cartId is null ");
+            throw new NotValidUsernameException("cartId is null ");
         }
 
         Cart cart = cartRepository.findCartByUserId(userId);
@@ -101,10 +104,11 @@ public class CartServiceImpl implements CartService {
      * @return cart without items which needed to be removed.
      */
     @Override
-    public Cart deleteFromCart(UUID cartId, List<UUID> itemIds) throws EntityNotFoundException, ValidationException {
+    public Cart deleteFromCart(UUID cartId, List<UUID> itemIds) throws EntityNotFoundException, NotValidUsernameException, NotAccessException {
 
         Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         String username;
+
         if (principal instanceof UserDetails) {
             username = ((UserDetails) principal).getUsername();
         } else {
@@ -113,7 +117,7 @@ public class CartServiceImpl implements CartService {
 
         if (cartId == null) {
             log.error("IN deleteFromCart - id is null ");
-            throw new ValidationException("cartId is null");
+            throw new NotValidUsernameException("cartId is null");
         }
 
         Cart dbCart = cartRepository.findCartByCartId(cartId);
@@ -123,15 +127,23 @@ public class CartServiceImpl implements CartService {
             throw new EntityNotFoundException("Cart doesn't exist ");
         }
 
-        Collection<Item> items = dbCart.getItems();
-        Collection<Item> toDelete = new ArrayList<>();
+        User user = userRepository.findUserByUserId(dbCart.getUserId());
 
-        for (Item item : items) {
-            if (itemIds.contains(item.getItemId()))
-                toDelete.add(item);
+        if (Objects.equals(username, user.getMail())) {
+            Collection<Item> items = dbCart.getItems();
+            Collection<Item> toDelete = new ArrayList<>();
+
+            for (Item item : items) {
+                if (itemIds.contains(item.getItemId()))
+                    toDelete.add(item);
+            }
+
+            items.removeAll(toDelete);
+        } else {
+            log.error("IN deleteFromCart - This username can't delete items from this cart");
+            throw new NotAccessException("Username is invalid");
         }
 
-        items.removeAll(toDelete);
         log.info("IN deleteFromCart - cart with id: {} successfully delete {} items", cartId, itemIds.size());
         return cartRepository.save(dbCart);
     }
@@ -167,11 +179,20 @@ public class CartServiceImpl implements CartService {
      * @return Updated cart.
      */
     @Override
-    public Cart updateCart(UUID cartId, Cart cart) throws ValidationException, EntityNotFoundException {
+    public Cart updateCart(UUID cartId, Cart cart) throws NotValidUsernameException, EntityNotFoundException, NotAccessException {
+
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username;
+
+        if (principal instanceof UserDetails) {
+            username = ((UserDetails) principal).getUsername();
+        } else {
+            username = principal.toString();
+        }
 
         if (cart == null) {
             log.error("IN updateCart - cart is null");
-            throw new ValidationException("cart is null");
+            throw new NotValidUsernameException("cart is null");
         }
 
         Cart dbCart = cartRepository.findCartByCartId(cartId);
@@ -181,19 +202,24 @@ public class CartServiceImpl implements CartService {
             throw new EntityNotFoundException("cart not found");
         }
 
-        Collection<Item> cartsItem = new LinkedList<>();
+        User user = userRepository.findUserByUserId(dbCart.getUserId());
 
-        if (cart.getItems() != null) {
-            for (Item item : cart.getItems()) {
-                Item savedItem = itemRepository.findItemByItemId(item.getItemId());
-                cartsItem.add(savedItem);
+        if (Objects.equals(username, user.getMail())) {
+            Collection<Item> cartsItem = new LinkedList<>();
+
+            if (cart.getItems() != null) {
+                for (Item item : cart.getItems()) {
+                    Item savedItem = itemRepository.findItemByItemId(item.getItemId());
+                    cartsItem.add(savedItem);
+                }
+                dbCart.setItems(cartsItem);
             }
-            dbCart.setItems(cartsItem);
+
+            log.info("IN updateCart - cart with id: {} successfully edited ", cartId);
+        } else {
+            log.error("IN updateCart - This username can't update this cart");
+            throw new NotAccessException("Username is invalid");
         }
-
-        dbCart.setUserId(cart.getUserId());
-
-        log.info("IN updateCart - cart with id: {} successfully edited ", cartId);
         return cartRepository.save(dbCart);
     }
 
@@ -203,8 +229,32 @@ public class CartServiceImpl implements CartService {
      * @param cartId This is cart's id which needed to delete.
      */
     @Override
-    public void deleteCartById(UUID cartId) {
-        cartRepository.deleteByCartId(cartId);
-        log.info("IN delete - cart with id: {} successfully deleted", cartId);
+    public void deleteCartById(UUID cartId) throws EntityNotFoundException, NotAccessException {
+
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username;
+
+        if (principal instanceof UserDetails) {
+            username = ((UserDetails) principal).getUsername();
+        } else {
+            username = principal.toString();
+        }
+
+        Cart dbCart = cartRepository.findCartByCartId(cartId);
+
+        if (dbCart == null) {
+            log.error("IN updateCart - no cart found by id: {}", cartId);
+            throw new EntityNotFoundException("cart not found");
+        }
+
+        User user = userRepository.findUserByUserId(dbCart.getUserId());
+
+        if (Objects.equals(username, user.getMail())) {
+            cartRepository.deleteByCartId(cartId);
+            log.info("IN delete - cart with id: {} successfully deleted", cartId);
+        } else {
+            log.error("IN updateCart - This username can't update this cart");
+            throw new NotAccessException("Username is invalid");
+        }
     }
 }
